+++ run.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/with-contenv bashio
 
 mode=$(bashio::config 'GSM_Mode')
-
 device=$(bashio::config 'GSM_Device')
 pin=$(bashio::config 'GSM_PIN')
 auth=$(bashio::config 'GSM_AUTH')
@@ -13,5 +12,7 @@ password=$(bashio::config 'MQTT_Password')
 send=$(bashio::config 'MQTT_Send')
 recv=$(bashio::config 'MQTT_Receive')
 
+logging=$(bashio::config 'ADDON_Logging')
+
 echo "run.sh: launching sms_manager.py"
-python3 /sms_manager.py --mode $mode -d $device --pin $pin --host $host --port $port -u $user -s $password --auth $auth --send $send --recv $recv
+python3 /sms_manager.py --mode $mode -d $device --pin $pin --host $host --port $port -u $user -s $password --auth $auth --send $send --recv $recv --log $logging